feat: Set timestep from double to float

+++ DXViewer-1.0.0/include/DX12App.h
@@ -13,7 +13,7 @@ public:
     __declspec(dllexport) DX12App();
     __declspec(dllexport) ~DX12App();
 
-    __declspec(dllexport) void setSimulation(ISimulation* simulation, double timestep);
+    __declspec(dllexport) void setSimulation(ISimulation* simulation, float timestep);
     void setProjectionType(PROJ proj);
 
     void setWindow(const int kWidth, const int kHeight, HWND mhMainWnd);
